Generate code from protos with serde support
This will be later useful for JWT implementation. One downside right now
is that `with-serde` feature has to be set as the default feature since
the API code is generated with `cfg_attr` which cannot be turned off
even it's stated otherwise in `protobuf` docs. I have filed a bug report
here: https://github.com/stepancheg/rust-protobuf/issues/516